smpc: moved share builder into its own file
